data_server/requirements.txt
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
